refactor: extract approved egress config overlay
